+++ CHANGELOG.md
@@ -29,6 +29,8 @@
 - Don't convert to multi-part geometries by default in `copy_layer`,... (#570)
 - Add configuration option to only warn on dissolve errors (#561)
 - Add some pre-flight checks when geofileops is imported (#573, #627)
+- For `select_two_layers`, add the `gpkg_ogr_contents` table + fill out extents in the
+  `gpkg_contents` table in the output file (#647)
 - When using join_nearest with spatialite version >= 5.1,
   show ST_distance between the two geometries instead of 
   the distance between the centroid of the two geometries (#634)